refactor: rename join_fanin module to join_queue across all adapters for consistency
