+++ aimdb-wasm-adapter/src/lib.rs
@@ -35,7 +35,7 @@
 extern crate alloc;
 
 pub mod buffer;
-pub mod join_fanin;
+pub mod join_queue;
 pub mod logger;
 pub mod runtime;
 pub mod time;
